Refactor:#51 페이지 컴포넌트의 비즈니스 로직을 커스텀 Hook으로 분리
FSD 컨벤션에 따라 pages에서 entities/features model 레이어로 추출:
- friend: useFriendList, useFriendRequests, useFriendSearch, usePendingRequestCount
- template: useMyTemplates, useTemplateForm
- wrong-note: useWrongNoteList, useWrongNoteDetail, useWrongNoteForm, useFriendWrongNotes
- grading: useGrading
